+++ app/lib/shared/versions.dart
@@ -22,7 +22,7 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2023.05.03',
+  '2023.05.04',
   // Fallback runtime versions.
   '2023.04.27',
   '2023.04.21',
@@ -54,9 +54,9 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '2.19.6';
-final String toolStableDartSdkVersion = '3.0.0-417.4.beta';
+final String toolStableDartSdkVersion = '3.0.0';
 final String toolStableFlutterSdkVersion = '3.10.0-1.4.pre';
-final String toolPreviewDartSdkVersion = '3.0.0-417.4.beta';
+final String toolPreviewDartSdkVersion = '3.0.0';
 final String toolPreviewFlutterSdkVersion = '3.10.0-1.4.pre';
 
 final semanticToolStableDartSdkVersion =