+++ CHANGELOG.md
@@ -2,9 +2,9 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
- * Bumped runtimeVersion to `2023.05.03`.
- * Upgraded stable Dart analysis SDK to `3.0.0-417.4.beta`.
- * Upgraded preview Dart analysis SDK to `3.0.0-417.4.beta`.
+ * Bumped runtimeVersion to `2023.05.04`.
+ * Upgraded stable Dart analysis SDK to `3.0.0`.
+ * Upgraded preview Dart analysis SDK to `3.0.0`.
  * Upgraded stable Flutter analysis SDK to `3.10.0-1.4.pre`.
  * Upgraded pana to `0.21.31`.
  * Upgraded dartdoc to `6.2.2`.